Derive Macro gd_props_macros::GdPropSaver

#[derive(GdPropSaver)]
{
    // Attributes available to this derive:
    #[register]
}

Create resource saver for GdProp resources.

Macro used to implement GdPropSaver trait for a bare rust-defined ResourceFormatSaver, allowing registered resources deriving GdProp to be saved using it to .gdron and .gdbin files.

Alongside implementing above trait, macro also implements IResourceFormatSaver, so you can’t implement it yourself.

Macro attributes

#[register(MyGdPropource, MyOtherGdPropource)] - registers Resources deriving GdProp to be handled by this struct. You can provide multiple resource in one attribute, you can also add multiple register attributes with resources.

Example

use godot::prelude::GodotClass;
use gd_props::GdPropSaver;

#[derive(GodotClass, GdPropSaver)]
#[class(init, tool, base=ResourceFormatSaver)]
#[register(MyResource, MyOtherResource)]
pub struct MyRonSaver {}

Register your GdPropSaver

To make the Saver recognizable by editor, remember to add tool value to the GodotClass macro #[class] attribute. Additionally, you need to register the saver in the ResourceSaver singleton at Godot runtime initialization. Recommended way of registration is to call GdPropourceSaver::register_saver() associated function in ExtensionLibrary implementation:

use godot::init::*;

struct MyGdExtension;

unsafe impl ExtensionLibrary for MyGdExtension {
    fn on_level_init(_level: InitLevel) {
        use gd_props::traits::GdPropSaver as _;
        if _level == InitLevel::Scene {
            MyResSaver::register_saver();
        }   
    }
}